Issue #35: Link to Poem and Poet
The daily poem now includes appropriate links to the poem and poet.
