+++ pages/_vertical/poet/index.vue
@@ -1,64 +1,64 @@
 <template>
   <div>
-    <b-container class="poems-list py-5">
-      <b-row class="poems-list__row">
+    <b-container class="poets-list py-5">
+      <b-row class="poets-list__row">
         <b-col md="4">
-          <div class="poems-list__list-header">
+          <div class="poets-list__list-header">
             Name
           </div>
         </b-col>
         <b-col md="4">
-          <div class="poems-list__list-header">
+          <div class="poets-list__list-header">
             Years
           </div>
         </b-col>
         <b-col md="4">
-          <div class="poems-list__list-header">
+          <div class="poets-list__list-header">
             Schools and Movements
           </div>
         </b-col>
       </b-row>
       <b-row
-        v-for="poem in poems"
-        class="poems-list__row poems-list__poems"
-        :key="poem.id">
+        v-for="poet in poets"
+        class="poets-list__row poets-list__poems"
+        :key="poet.id">
         <b-col md="4">
-          <a :href="`/poems/${getPoemTitle(poem.view_node)}`">
-            {{ poem.title }}
+          <a :href="`/poets/${getPoetTitle(poet.view_node)}`">
+            {{ poet.poets }}
           </a>
         </b-col>
         <b-col md="4">
-          {{ poem.nothing }}
+          {{ poet.nothing }}
         </b-col>
         <b-col md="4">
-          {{ poem.field_school_movement }}
+          {{ poet.field_school_movement }}
         </b-col>
       </b-row>
       <b-row>
         <b-col md="4">
-          <a :href="`/poems?page=${Prev}`">
+          <a :href="`/poets?page=${Prev}`">
             &lt;&lt; Prev
           </a>
         </b-col>
         <b-col md="4">
-          <a href="/poems?page=1">
+          <a href="/poets?page=1">
             1
           </a>
-          <a href="/poems?page=2">
+          <a href="/poets?page=2">
             2
           </a>
-          <a href="/poems?page=3">
+          <a href="/poets?page=3">
             3
           </a>
           . . .
-          <a :href="`/poems?page=${$store.state.numPages.numPages}`">
+          <a :href="`/poets?page=${$store.state.numPages.numPages}`">
             {{ $store.state.numPages.numPages }}
           </a>
         </b-col>
         <b-col md="4">
           <a
             v-if="$store.state.numPages.numPages > Next"
-            :href="`/poems?page=${Next}`">
+            :href="`/poets?page=${Next}`">
             Next &gt;&gt;
           </a>
         </b-col>
@@ -68,22 +68,27 @@
 </template>
 
 <script>
-import AppPoemADaySignUpForm from "../../components/AppPoemADayPoems/AppPoemADaySignUpForm";
-import AppPoems from "../../components/AppPoemADayPoems/AppPoems";
-import paginationHelpers from "../../plugins/pagination-helpers";
+import AppPoemADaySignUpForm from "~/components/AppPoemADayPoems/AppPoemADaySignUpForm";
+import AppPoems from "~/components/AppPoemADayPoems/AppPoems";
+import PoetList from "~/components/PoetList";
+import paginationHelpers from "~/plugins/pagination-helpers";
 export default {
-  components: { AppPoemADaySignUpForm, AppPoems },
+  components: {
+    AppPoemADaySignUpForm,
+    AppPoems,
+    PoetList
+  },
   async asyncData({ app, params, query }) {
     const pageLinks = paginationHelpers.getPageLinks(query);
     return app.$axios
-      .get(`/api/poems`, {
+      .get(`/api/poets`, {
         params: {
           page: pageLinks.pageNum
         }
       })
       .then(res => {
         return {
-          poems: res.data,
+          poets: res.data,
           pageNum: pageLinks.pageNum,
           Prev: pageLinks.Prev,
           Next: pageLinks.Next
@@ -96,13 +101,13 @@ export default {
   async fetch({ app, store, params }) {
     const numPages = await paginationHelpers.getNumPages(
       app,
-      "/api/node/poems",
-      50
+      "/api/node/person",
+      20
     );
     store.commit("updateNumPages", numPages);
   },
   methods: {
-    getPoemTitle(viewNode) {
+    getPoetTitle(viewNode) {
       const title = viewNode.split("/");
       return title[3];
     }
@@ -111,15 +116,15 @@ export default {
 </script>
 
 <style scoped lang="scss">
-.poems-list__row {
+.poets-list__row {
   border-bottom: 1px solid #ccc;
   font-size: 18px;
   min-height: 47px;
 }
-.poems-list__list-header {
+.poets-list__list-header {
   font-weight: bold;
 }
-.poems-list__poems {
+.poets-list__poems {
   font-weight: 111;
 }
 </style>
